Extract VideoPlayButton as a separate widget.
