+++ lib/ui/onboarding2/Onboarding2VideoTutorialPanel.dart
@@ -24,6 +24,7 @@ import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/NativeCommunicator.dart';
 import 'package:illinois/ui/onboarding2/Onboadring2RolesPanel.dart';
 import 'package:illinois/ui/onboarding2/Onboarding2Widgets.dart';
+import 'package:illinois/ui/widgets/VideoPlayButton.dart';
 import 'package:rokwire_plugin/service/app_navigation.dart';
 import 'package:rokwire_plugin/service/assets.dart';
 import 'package:rokwire_plugin/service/config.dart';
@@ -32,7 +33,6 @@ import 'package:rokwire_plugin/service/network.dart';
 import 'package:rokwire_plugin/service/notification_service.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/ui/widgets/rounded_button.dart';
-import 'package:rokwire_plugin/ui/widgets/triangle_painter.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
 import 'package:video_player/video_player.dart';
 
@@ -191,7 +191,7 @@ class _Onboarding2VideoTutorialPanelState extends State<Onboarding2VideoTutorial
                           ClosedCaption(
                               text: _currentCaptionText, textStyle: TextStyle(fontSize: 16, color: Styles().colors!.white))
                         ]),
-                        _buildPlayButton()
+                        Visibility(visible: (_isPlayerInitialized && !_isPlaying), child: VideoPlayButton())
                       ]))));
             } else {
               return const Center(child: CircularProgressIndicator());
@@ -202,37 +202,6 @@ class _Onboarding2VideoTutorialPanelState extends State<Onboarding2VideoTutorial
           child: Text(Localization().getStringEx('panel.onboarding2.video.missing.msg', 'Missing video'),
               style: TextStyle(color: Styles().colors!.white, fontSize: 20, fontFamily: Styles().fontFamilies!.bold)));
     }
-  }
-
-  Widget _buildPlayButton() {
-    final double buttonWidth = 80;
-    final double buttonHeight = 50;
-    bool buttonVisible = _isPlayerInitialized && !_isPlaying;
-    return Visibility(
-        visible: buttonVisible,
-        child: Container(
-            decoration: BoxDecoration(color: Styles().colors!.iconColor, borderRadius: BorderRadius.all(Radius.circular(10))),
-            width: buttonWidth,
-            height: buttonHeight,
-            child: Center(
-                child: Column(mainAxisAlignment: MainAxisAlignment.center, children: [
-              Container(
-                  width: (buttonHeight / 2),
-                  child: CustomPaint(
-                      painter: TrianglePainter(
-                          painterColor: Styles().colors!.white,
-                          horzDir: TriangleHorzDirection.rightToLeft,
-                          vertDir: TriangleVertDirection.topToBottom),
-                      child: Container(height: (buttonHeight / 4)))),
-              Container(
-                  width: (buttonHeight / 2),
-                  child: CustomPaint(
-                      painter: TrianglePainter(
-                          painterColor: Styles().colors!.white,
-                          horzDir: TriangleHorzDirection.rightToLeft,
-                          vertDir: TriangleVertDirection.bottomToTop),
-                      child: Container(height: (buttonHeight / 4))))
-            ]))));
   }
 
   Widget _buildCcButton() {
